// lib/api/types.ts
export interface ReportEntry {
  date?: string
  month?: string
  week?: string
  income?: number
  outcome?: number
  revenue?: number
  expenses?: number
  label?: string
  value?: number
}